make SType public,
split types module into submodules;
