+++ ergo-lib/src/serialization/types.rs
@@ -4,7 +4,7 @@ use super::sigma_byte_writer::SigmaByteWrite;
 use crate::serialization::{
     sigma_byte_reader::SigmaByteRead, SerializationError, SigmaSerializable,
 };
-use crate::types::SType;
+use crate::types::stype::SType;
 use sigma_ser::vlq_encode;
 use std::{io, ops::Add};
 use vlq_encode::WriteSigmaVlqExt;
@@ -92,6 +92,17 @@ fn is_stype_embeddable(tpe: &SType) -> bool {
     }
 }
 
+/**
+ * Each SType is serialized to array of bytes by:
+ * - emitting typeCode of each node (see special case for collections below)
+ * - then recursively serializing subtrees from left to right on each level
+ * - for each collection of primitive type there is special type code to emit single byte instead of two bytes
+ * Types code intervals
+ * - (1 .. MaxPrimTypeCode)  // primitive types
+ * - (CollectionTypeCode .. CollectionTypeCode + MaxPrimTypeCode) // collections of primitive types
+ * - (MaxCollectionTypeCode ..)  // Other types
+ * Collection of non-primitive type is serialized as (CollectionTypeCode, serialize(elementType))
+ */
 impl SigmaSerializable for SType {
     fn sigma_serialize<W: SigmaByteWrite>(&self, w: &mut W) -> Result<(), io::Error> {
         // for reference see http://github.com/ScorexFoundation/sigmastate-interpreter/blob/25251c1313b0131835f92099f02cef8a5d932b5e/sigmastate/src/main/scala/sigmastate/serialization/TypeSerializer.scala#L25-L25
