+++ ergo-lib/src/sigma_protocol/verifier.rs
@@ -121,13 +121,11 @@ mod tests {
     use super::*;
     use crate::ast::constant::Constant;
     use crate::ast::expr::Expr;
-    use crate::{
-        sigma_protocol::{
-            private_input::{DlogProverInput, PrivateInput},
-            prover::{Prover, TestProver},
-        },
-        types::SType,
+    use crate::sigma_protocol::{
+        private_input::{DlogProverInput, PrivateInput},
+        prover::{Prover, TestProver},
     };
+    use crate::types::stype::SType;
     use proptest::prelude::*;
     use std::rc::Rc;
