+++ ergo-lib/src/types/sfunc.rs
@@ -1,0 +1,9 @@
+use super::stype::SType;
+use super::stype_param::STypeParam;
+
+#[derive(PartialEq, Eq, Debug, Clone)]
+pub struct SFunc {
+    t_dom: Vec<SType>,
+    t_range: SType,
+    tpe_params: Vec<STypeParam>,
+}
