+++ ergo-lib/src/types/stype_companion.rs
@@ -1,0 +1,11 @@
+use super::smethod::SMethod;
+
+#[derive(PartialEq, Eq, Debug, Clone)]
+pub struct TypeId(u8);
+
+#[derive(PartialEq, Eq, Debug, Clone)]
+pub struct STypeCompanion {
+    pub type_id: TypeId,
+    pub type_name: String,
+    pub methods: Vec<SMethod>,
+}
